// resources/js/types/models/others.ts
export interface Quiz {
    id: number;
    title: string;
    slug: string;
    description?: string | null;

    // Paramètres pédagogiques
    quiz_type: QuizType;
    time_limit_minutes?: number | null;
    max_attempts: number;
    is_mandatory: boolean;

    // Comportement
    shuffle_questions: boolean;
    shuffle_answers: boolean;
    show_results_immediately: boolean;
    show_correct_answers: boolean;

    // Scoring
    max_score?: number | null;
    success_threshold?: number | null;
    quize_instructions?: string | null;
    note_unit?: NoteUnit;
    // Publication
    status: SaveStatus;
    published_at?: string | null; // ISO date string
    created_at?: string;
    updated_at?: string;
    // Versioning
    version: number;
    questions?: QuizQuestion[] | [];
    activity?: CourseActivity | null;
}

export interface QuizQuestion {
    id: number;
    quiz_id: number;
    skill_id?: number;

    // Contenu
    question_text: string;
    question_type: QuizQuestionType;

    // Paramètres
    points: number;
    order: number;
    is_mandatory: boolean;
    quize?: Quiz;
    skill?: Skill;

    // Feedback
    feedback_correct?: string | null;
    feedback_incorrect?: string | null;
    answers?: QuizAnswer[] | [];
    created_at?: string;
    updated_at?: string;
}

export interface QuizAnswer {
    id: number;
    quiz_question_id: number;

    // Contenu
    answer_text?: string | null;
    is_correct: boolean;
    order: number;

    question?: QuizQuestion;
    // Numérique
    numeric_value?: number | null;
    tolerance?: number | null;
    created_at?: string;
    updated_at?: string;
}